+++ zoomimage-compose-coil3-core/src/commonTest/kotlin/com/github/panpf/zoomimage/compose/coil3/core/test/internal/EngineCoilComposeSubsamplingImageGeneratorTest.kt
@@ -31,7 +31,7 @@ class EngineCoilComposeSubsamplingImageGeneratorTest {
                     imageSource = CoilHttpImageSource.Factory(
                         context,
                         imageLoader,
-                        "http://sample.com/sample.jpeg"
+                        ImageRequest.Builder(context).data("http://sample.com/sample.jpeg").build()
                     ),
                     imageInfo = null
                 )