+++ zoomimage-core-coil3/src/commonMain/kotlin/com/github/panpf/zoomimage/coil/internal/coil_core_utils.common.kt
@@ -18,6 +18,7 @@ package com.github.panpf.zoomimage.coil.internal
 
 import coil3.ImageLoader
 import coil3.PlatformContext
+import coil3.request.ImageRequest
 import com.github.panpf.zoomimage.subsampling.ImageSource
 
 /**
@@ -26,6 +27,19 @@ import com.github.panpf.zoomimage.subsampling.ImageSource
  * @see com.github.panpf.zoomimage.core.coil3.ios.test.internal.CoilCoreUtilsIosTest.testDataToImageSource
  * @see com.github.panpf.zoomimage.core.coil3.jscommon.test.internal.CoilCoreUtilsJsCommonTest.testDataToImageSource
  */
+expect suspend fun dataToImageSource(
+    context: PlatformContext,
+    imageLoader: ImageLoader,
+    request: ImageRequest,
+): ImageSource.Factory?
+
+/**
+ * @see com.github.panpf.zoomimage.core.coil3.android.test.internal.CoilCoreUtilsAndroidTest.testDataToImageSource
+ * @see com.github.panpf.zoomimage.core.coil3.desktop.test.internal.CoilCoreUtilsDesktopTest.testDataToImageSource
+ * @see com.github.panpf.zoomimage.core.coil3.ios.test.internal.CoilCoreUtilsIosTest.testDataToImageSource
+ * @see com.github.panpf.zoomimage.core.coil3.jscommon.test.internal.CoilCoreUtilsJsCommonTest.testDataToImageSource
+ */
+@Deprecated("Please use dataToImageSource(context, imageLoader, request) instead")
 expect suspend fun dataToImageSource(
     context: PlatformContext,
     imageLoader: ImageLoader,