+++ zoomimage-view-coil3-core/src/androidTest/kotlin/com/github/panpf/zoomimage/view/coil3/core/test/internal/EngineCoilViewSubsamplingImageGeneratorTest.kt
@@ -31,7 +31,7 @@ class EngineCoilViewSubsamplingImageGeneratorTest {
                     imageSource = CoilHttpImageSource.Factory(
                         context,
                         imageLoader,
-                        "http://sample.com/sample.jpeg"
+                        ImageRequest.Builder(context).data("http://sample.com/sample.jpeg").build()
                     ),
                     imageInfo = null
                 )